doctor_patient_app/requirements.txt
# Dependencies for Doctor-Patient Real-Time Voice Translator

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0

# WebSocket & Networking
websockets==12.0
aiohttp==3.9.1

# Data Models & Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# Audio Processing
numpy==1.26.2

# LangChain (for token routing and processing)
langchain==0.1.0
langchain-core>=0.1.7,<0.2

# Utilities
typing-extensions==4.8.0
